separating gl.h/cpp into components
