+++ src/cinder/gl/draw.cpp
@@ -1,810 +1,37 @@
-#include "cinder/gl/gl.h"
-#include "cinder/gl/Context.h"
-#include "cinder/gl/VboMesh.h"
-#include "cinder/gl/Vao.h"
-#include "cinder/gl/Vbo.h"
-#include "cinder/gl/Texture.h"
-#include "cinder/gl/Batch.h"
-#include "cinder/gl/Shader.h"
-#include "cinder/gl/Environment.h"
-#include "cinder/gl/Fbo.h"
+/*
+ Copyright (c) 2015, The Cinder Project, All rights reserved.
 
-#include "cinder/Utilities.h"
+ This code is intended for use with the Cinder C++ library: http://libcinder.org
+
+ Redistribution and use in source and binary forms, with or without modification, are permitted provided that
+ the following conditions are met:
+
+ * Redistributions of source code must retain the above copyright notice, this list of conditions and
+	the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright notice, this list of conditions and
+	the following disclaimer in the documentation and/or other materials provided with the distribution.
+
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
+ WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+ ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "cinder/gl/draw.h"
+#include "cinder/gl/Context.h"
+#include "cinder/gl/Vao.h"
+#include "cinder/gl/VboMesh.h"
+#include "cinder/Log.h"
 #include "cinder/Text.h"
 #include "cinder/Triangulate.h"
-#include "cinder/TriMesh.h"
-#include "cinder/Path2d.h"
-#include "cinder/Shape2d.h"
-#include "cinder/Log.h"
-#include <vector>
-
-#if defined( CINDER_MSW )
-	#include "glload/wgl_all.h"
-#elif defined( CINDER_MAC )
-	#include <OpenGL/OpenGL.h>
-#endif
 
 using namespace std;
 
 namespace cinder { namespace gl {
-
-Context* context()
-{
-	return Context::getCurrent();
-}
-
-void enableVerticalSync( bool enable )
-{
-#if defined( CINDER_MAC )
-	GLint sync = ( enable ) ? 1 : 0;
-	::CGLSetParameter( ::CGLGetCurrentContext(), kCGLCPSwapInterval, &sync );
-#elif defined( CINDER_MSW ) && ! defined( CINDER_GL_ANGLE )
-	GLint sync = ( enable ) ? 1 : 0;
-	if( wglext_EXT_swap_control )
-		::wglSwapIntervalEXT( sync );
-#endif
-}
-
-bool isVerticalSyncEnabled()
-{
-#if defined( CINDER_MAC )
-	GLint enabled;
-	::CGLGetParameter( ::CGLGetCurrentContext(), kCGLCPSwapInterval, &enabled );
-	return enabled > 0;
-#elif defined( CINDER_MSW ) && ! defined( CINDER_GL_ANGLE )
-	if( wglext_EXT_swap_control )
-		return ::wglGetSwapIntervalEXT() > 0;
-	else
-		return true;
-#else
-	return true;
-#endif
-}
-
-bool isExtensionAvailable( const std::string &extName )
-{
-	return env()->isExtensionAvailable( extName );
-}
-
-std::pair<GLint,GLint> getVersion()
-{
-	//hard-coded for now
-#if defined( CINDER_GL_ES_2 )
-	return std::make_pair( (GLint)2, (GLint)0 );
-#elif defined( CINDER_GL_ES_3 )
-	return std::make_pair( (GLint)3, (GLint)0 );
-#else
-	static bool	sInitialized = false;
-	static pair<GLint,GLint> sVersion;
-	if( ! sInitialized ) {
-		// adapted from LoadOGL
-		const char *strVersion = reinterpret_cast<const char*>( glGetString( GL_VERSION ) );
-		GLint major = 0, minor = 0;
-		const char *strDotPos = NULL;
-		int iLength = 0;
-		char strWorkBuff[10];
-
-		strDotPos = strchr( strVersion, '.' );
-		if( ! strDotPos )
-			return std::make_pair( 0, 0 );
-
-		iLength = (int)((ptrdiff_t)strDotPos - (ptrdiff_t)strVersion);
-		strncpy(strWorkBuff, strVersion, iLength);
-		strWorkBuff[iLength] = '\0';
-
-		major = atoi(strWorkBuff);
-		strDotPos = strchr( strVersion + iLength + 1, ' ' );
-		if( ! strDotPos ) { // No extra data. Take the whole rest of the string.
-			strcpy( strWorkBuff, strVersion + iLength + 1 );
-		}
-		else {
-			// Copy only up until the space.
-			int iLengthMinor = (int)((ptrdiff_t)strDotPos - (ptrdiff_t)strVersion);
-			iLengthMinor = iLengthMinor - (iLength + 1);
-			strncpy( strWorkBuff, strVersion + iLength + 1, iLengthMinor );
-			strWorkBuff[iLengthMinor] = '\0';
-		}
-
-		minor = atoi( strWorkBuff );
-		sVersion = std::make_pair( major, minor );
-		sInitialized = true;
-	}
-
-	return sVersion;
-#endif
-}
-
-std::string getVersionString()
-{
-	const GLubyte* s = glGetString( GL_VERSION );
-
-	return std::string( reinterpret_cast<const char*>( s ) );
-}
-
-GlslProgRef& getStockShader( const class ShaderDef &shader )
-{
-	return context()->getStockShader( shader );
-}
-
-void bindStockShader( const class ShaderDef &shaderDef )
-{
-	auto ctx = gl::context();
-	auto shader = ctx->getStockShader( shaderDef );
-	ctx->bindGlslProg( shader );
-}
-
-void setDefaultShaderVars()
-{
-	auto ctx = gl::context();
-	ctx->setDefaultShaderVars();
-}
-
-void clear( const ColorA& color, bool clearDepthBuffer )
-{
-	clearColor( color );
-	if ( clearDepthBuffer ) {
-		depthMask( GL_TRUE );
-		clear( GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT );
-	} else {
-		clear( GL_COLOR_BUFFER_BIT );
-	}
-}
-
-void clear( GLbitfield mask )
-{
-    glClear( mask );
-}
-
-void clearColor( const ColorA &color )
-{
-    glClearColor( color.r, color.g, color.b, color.a );
-}
-
-void clearDepth( const double depth )
-{
-#if ! defined( CINDER_GL_ES )
-    glClearDepth( depth );
-#else
-	glClearDepthf( depth );
-#endif
-}
-
-void clearDepth( const float depth )
-{
-    glClearDepthf( depth );
-}
-
-void clearStencil( const int s )
-{
-    glClearStencil( s );
-}
-
-void colorMask( GLboolean red, GLboolean green, GLboolean blue, GLboolean alpha )
-{
-    glColorMask( red, green, blue, alpha );
-}
-
-void depthMask( GLboolean flag )
-{
-    auto ctx = gl::context();
-	ctx->depthMask( flag );
-}
-
-void stencilMask( GLboolean mask )
-{
-	glStencilMask( mask );
-}
-
-void stencilFunc( GLenum func, GLint ref, GLuint mask )
-{
-    glStencilFunc( func, ref, mask );
-}
-
-void stencilOp( GLenum fail, GLenum zfail, GLenum zpass )
-{
-    glStencilOp( fail, zfail, zpass );
-}
-
-std::pair<ivec2, ivec2> getViewport()
-{
-	auto ctx = gl::context();
-	auto view = ctx->getViewport();
-	return view;
-}
-
-void viewport( const std::pair<ivec2, ivec2> positionAndSize )
-{
-	auto ctx = gl::context();
-	ctx->viewport( positionAndSize );
-}
-
-void pushViewport( const std::pair<ivec2, ivec2> positionAndSize )
-{
-	auto ctx = gl::context();
-	ctx->pushViewport( positionAndSize );
-}
-
-void popViewport()
-{
-	auto ctx = gl::context();
-	ctx->popViewport();
-}
-
-std::pair<ivec2, ivec2> getScissor()
-{
-	auto ctx = gl::context();
-	auto scissor = ctx->getScissor();
-	return scissor;
-}
-
-void scissor( const std::pair<ivec2, ivec2> positionAndSize )
-{
-	auto ctx = gl::context();
-	ctx->setScissor( positionAndSize );
-}
-
-void enable( GLenum state, bool enable )
-{
-	auto ctx = gl::context();
-	ctx->enable( state, enable );
-}
-
-void enableAlphaBlending( bool premultiplied )
-{
-	auto ctx = gl::context();
-	ctx->enable( GL_BLEND );
-	if( ! premultiplied ) {
-		ctx->blendFunc( GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA );
-	}
-	else {
-		ctx->blendFunc( GL_ONE, GL_ONE_MINUS_SRC_ALPHA );
-	}
-}
-
-void disableAlphaBlending()
-{
-	gl::disable( GL_BLEND );
-}
-
-void enableAdditiveBlending()
-{
-	auto ctx = gl::context();
-	ctx->enable( GL_BLEND );
-	ctx->blendFunc( GL_SRC_ALPHA, GL_ONE );
-}
-
-void enableFaceCulling( bool enable )
-{
-	gl::context()->setBoolState( GL_CULL_FACE, enable );
-}
-
-void cullFace( GLenum face )
-{
-	gl::context()->cullFace( face );
-}
-
-void disableDepthRead()
-{
-	gl::disable( GL_DEPTH_TEST );
-}
-
-void enableDepthRead( bool enable )
-{
-	gl::enable( GL_DEPTH_TEST, enable );
-}
-
-void enableDepthWrite( bool enable )
-{
-	auto ctx = gl::context();
-	ctx->depthMask( enable ? GL_TRUE : GL_FALSE );
-}
-
-void disableDepthWrite()
-{
-	auto ctx = gl::context();
-	ctx->depthMask( GL_FALSE );
-}
-
-void enableStencilTest( bool enable )
-{
-    gl::enable( GL_STENCIL_TEST, enable );
-}
-
-void disableStencilTest()
-{
-    gl::disable( GL_STENCIL_TEST );
-}
-
-void setMatrices( const ci::Camera& cam )
-{
-	auto ctx = context();
-	ctx->getViewMatrixStack().back() = cam.getViewMatrix();
-	ctx->getProjectionMatrixStack().back() = cam.getProjectionMatrix();
-	ctx->getModelMatrixStack().back() = mat4();
-}
-
-void setModelMatrix( const ci::mat4 &m )
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().back() = m;
-}
-
-void setViewMatrix( const ci::mat4 &m )
-{
-	auto ctx = context();
-	ctx->getViewMatrixStack().back() = m;
-}
-
-void setProjectionMatrix( const ci::mat4 &m )
-{
-	auto ctx = context();
-	ctx->getProjectionMatrixStack().back() = m;
-}
-
-void pushModelMatrix()
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().push_back( ctx->getModelMatrixStack().back() );
-}
-
-void popModelMatrix()
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().pop_back();
-}
-
-void pushViewMatrix()
-{
-	auto ctx = context();
-	ctx->getViewMatrixStack().push_back( ctx->getViewMatrixStack().back() );
-}
-
-void popViewMatrix()
-{
-	auto ctx = context();
-	ctx->getViewMatrixStack().pop_back();
-}
-
-void pushProjectionMatrix()
-{
-	auto ctx = context();
-	ctx->getProjectionMatrixStack().push_back( ctx->getProjectionMatrixStack().back() );
-}
-
-void popProjectionMatrix()
-{
-	auto ctx = context();
-	ctx->getProjectionMatrixStack().pop_back();
-}
-
-void pushModelView()
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().push_back( ctx->getModelMatrixStack().back() );
-	ctx->getViewMatrixStack().push_back( ctx->getViewMatrixStack().back() );
-}
-
-void popModelView()
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().pop_back();
-	ctx->getViewMatrixStack().pop_back();
-}
-
-void pushMatrices()
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().push_back( ctx->getModelMatrixStack().back() );
-	ctx->getViewMatrixStack().push_back( ctx->getViewMatrixStack().back() );
-	ctx->getProjectionMatrixStack().push_back( ctx->getProjectionMatrixStack().back() );
-}
-
-void popMatrices()
-{
-	auto ctx = context();
-	ctx->getModelMatrixStack().pop_back();
-	ctx->getViewMatrixStack().pop_back();
-	ctx->getProjectionMatrixStack().pop_back();
-}
-
-void multModelMatrix( const ci::mat4& mtx )
-{
-	auto ctx = gl::context();
-	ctx->getModelMatrixStack().back() *= mtx;
-}
-
-void multViewMatrix( const ci::mat4& mtx )
-{
-	auto ctx = gl::context();
-	ctx->getViewMatrixStack().back() *= mtx;
-}
-
-void multProjectionMatrix( const ci::mat4& mtx )
-{
-	auto ctx = gl::context();
-	ctx->getProjectionMatrixStack().back() *= mtx;
-}
-
-mat4 getModelMatrix()
-{
-	auto ctx = gl::context();
-	return ctx->getModelMatrixStack().back();
-}
-
-mat4 getViewMatrix()
-{
-	auto ctx = gl::context();
-	return ctx->getViewMatrixStack().back();
-}
-
-mat4 getProjectionMatrix()
-{
-	auto ctx = gl::context();
-	return ctx->getProjectionMatrixStack().back();
-}
-
-mat4 getModelView()
-{
-	auto ctx = context();
-	return ctx->getViewMatrixStack().back() * ctx->getModelMatrixStack().back();
-}
-
-mat4 getModelViewProjection()
-{
-	auto ctx = context();
-	return ctx->getProjectionMatrixStack().back() * ctx->getViewMatrixStack().back() * ctx->getModelMatrixStack().back();
-}
-
-mat4 calcViewMatrixInverse()
-{
-	return glm::inverse( getViewMatrix() );
-}
-
-mat3 calcNormalMatrix()
-{
-	return glm::inverseTranspose( glm::mat3( getModelView() ) );
-}
-	
-mat3 calcModelMatrixInverseTranspose()
-{
-	auto m = glm::inverseTranspose( getModelMatrix() );
-	return mat3( m );
-}
-	
-mat4 calcViewportMatrix()
-{
-	auto curViewport = getViewport();
-	
-	const float a = ( curViewport.second.x - curViewport.first.x ) / 2.0f;
-	const float b = ( curViewport.second.y - curViewport.first.y ) / 2.0f;
-	const float c = 1.0f / 2.0f;
-	
-	const float tx = ( curViewport.second.x + curViewport.first.x ) / 2.0f;
-	const float ty = ( curViewport.second.y + curViewport.second.y ) / 2.0f;
-	const float tz = 1.0f / 2.0f;
-	
-	return mat4(
-		a, 0, 0, 0,
-		0, b, 0, 0,
-		0, 0, c, 0,
-		tx, ty, tz, 1
-	);
-}
-
-void setMatricesWindowPersp( int screenWidth, int screenHeight, float fovDegrees, float nearPlane, float farPlane, bool originUpperLeft )
-{
-	auto ctx = gl::context();
-
-	CameraPersp cam( screenWidth, screenHeight, fovDegrees, nearPlane, farPlane );
-	ctx->getModelMatrixStack().back() = mat4();
-	ctx->getProjectionMatrixStack().back() = cam.getProjectionMatrix();
-	ctx->getViewMatrixStack().back() = cam.getViewMatrix();
-	if( originUpperLeft ) {
-		ctx->getViewMatrixStack().back() *= glm::scale( vec3( 1, -1, 1 ) );								// invert Y axis so increasing Y goes down.
-		ctx->getViewMatrixStack().back() *= glm::translate( vec3( 0, (float) - screenHeight, 0 ) );		// shift origin up to upper-left corner.
-	}
-}
-
-void setMatricesWindowPersp( const ci::ivec2& screenSize, float fovDegrees, float nearPlane, float farPlane, bool originUpperLeft )
-{
-	setMatricesWindowPersp( screenSize.x, screenSize.y, fovDegrees, nearPlane, farPlane, originUpperLeft );
-}
-
-void setMatricesWindow( int screenWidth, int screenHeight, bool originUpperLeft )
-{
-	auto ctx = gl::context();
-	ctx->getModelMatrixStack().back() = mat4();
-	ctx->getViewMatrixStack().back() = mat4();
-
-	float sx = 2.0f / (float)screenWidth;
-	float sy = 2.0f / (float)screenHeight;
-	float ty = -1;
-
-	if( originUpperLeft ) {
-		sy *= -1;
-		ty *= -1;
-	}
-
-	mat4 &m = ctx->getProjectionMatrixStack().back();
-	m = mat4( sx, 0,  0, 0,
-			  0, sy,  0, 0,
-			  0,  0, -1, 0,
-			 -1, ty,  0, 1 );
-}
-
-void setMatricesWindow( const ci::ivec2& screenSize, bool originUpperLeft )
-{
-	setMatricesWindow( screenSize.x, screenSize.y, originUpperLeft );
-}
-
-void rotate( const quat &quat )
-{
-	auto ctx = gl::context();
-	ctx->getModelMatrixStack().back() *= toMat4( quat );
-}
-
-void rotate( float angleRadians, const vec3 &axis )
-{
-	if( math<float>::abs( angleRadians ) > EPSILON_VALUE ) {
-		auto ctx = gl::context();
-		ctx->getModelMatrixStack().back() *= glm::rotate( angleRadians, axis );
-	}
-}
-
-void scale( const ci::vec3& v )
-{
-	auto ctx = gl::context();
-	ctx->getModelMatrixStack().back() *= glm::scale( v );
-}
-
-void translate( const ci::vec3& v )
-{
-	auto ctx = gl::context();
-	ctx->getModelMatrixStack().back() *= glm::translate( v );
-}
-
-vec3 windowToObjectCoord( const mat4 &modelMatrix, const ci::vec2 &coordinate, float z )
-{
-	// Build the viewport (x, y, width, height).
-	vec2 offset = gl::getViewport().first;
-	vec2 size = gl::getViewport().second;
-	vec4 viewport = vec4( offset.x, offset.y, size.x, size.y );
-
-	// Calculate the view-projection matrix.
-	mat4 viewProjectionMatrix = gl::getProjectionMatrix() * gl::getViewMatrix();
-
-	// Calculate the intersection of the mouse ray with the near (z=0) and far (z=1) planes.
-	vec3 nearPlane = glm::unProject( vec3( coordinate.x, size.y - coordinate.y, 0 ), modelMatrix, viewProjectionMatrix, viewport );
-	vec3 farPlane = glm::unProject( vec3( coordinate.x, size.y - coordinate.y, 1 ), modelMatrix, viewProjectionMatrix, viewport );
-
-	// Calculate world position.
-	return ci::lerp( nearPlane, farPlane, ( z - nearPlane.z ) / ( farPlane.z - nearPlane.z ) );
-}
-
-vec2 objectToWindowCoord( const mat4 &modelMatrix, const ci::vec3 &coordinate )
-{
-	// Build the viewport (x, y, width, height).
-	vec2 offset = gl::getViewport().first;
-	vec2 size = gl::getViewport().second;
-	vec4 viewport = vec4( offset.x, offset.y, size.x, size.y );
-
-	// Calculate the view-projection matrix.
-	mat4 viewProjectionMatrix = gl::getProjectionMatrix() * gl::getViewMatrix();
-
-	vec2 p = vec2( glm::project( coordinate, modelMatrix, viewProjectionMatrix, viewport ) );
-
-	return p;
-}
-
-void begin( GLenum mode )
-{
-	auto ctx = gl::context();
-	ctx->immediate().begin( mode );
-}
-
-void end()
-{
-	auto ctx = gl::context();
-
-	if( ctx->immediate().empty() )
-		return;
-	else {
-		ScopedGlslProg ScopedGlslProg( ctx->getStockShader( ShaderDef().color() ) );
-		ctx->immediate().draw();
-		ctx->immediate().clear();
-	}
-}
-
-#if ! defined( CINDER_GL_ES_2 )
-void bindBufferBase( GLenum target, int index, BufferObjRef buffer )
-{
-	auto ctx = gl::context();
-	ctx->bindBufferBase( target, index, buffer );
-}
-
-void beginTransformFeedback( GLenum primitiveMode )
-{
-	auto ctx = gl::context();
-	ctx->beginTransformFeedback( primitiveMode );
-}
-
-void pauseTransformFeedback()
-{
-	auto ctx = gl::context();
-	ctx->pauseTransformFeedback();
-}
-
-void resumeTransformFeedback()
-{
-	auto ctx = gl::context();
-	ctx->resumeTransformFeedback();
-}
-
-void endTransformFeedback()
-{
-	auto ctx = gl::context();
-	ctx->endTransformFeedback();
-}
-#endif // ! defined( CINDER_GL_ES_2 )
-
-#if ! defined( CINDER_GL_ES )
-void patchParameteri( GLenum pname, GLint value )
-{
-	glPatchParameteri( pname, value );
-}
-
-void patchParameterfv( GLenum pname, GLfloat *value )
-{
-	glPatchParameterfv( pname, value );
-}
-#endif // ! defined( CINDER_GL_ES )
-
-void color( float r, float g, float b )
-{
-	auto ctx = gl::context();
-	ctx->setCurrentColor( ColorAf( r, g, b, 1.0f ) );
-}
-
-void color( float r, float g, float b, float a )
-{
-	auto ctx = gl::context();
-	ctx->setCurrentColor( ColorAf( r, g, b, a ) );
-}
-
-void color( const ci::Color &c )
-{
-	auto ctx = gl::context();
-	ctx->setCurrentColor( c );
-}
-
-void color( const ci::ColorA &c )
-{
-	auto ctx = gl::context();
-	ctx->setCurrentColor( c );
-}
-
-void color( const ci::Color8u &c )
-{
-	auto ctx = gl::context();
-	ctx->setCurrentColor( c );
-}
-
-void color( const ci::ColorA8u &c )
-{
-	auto ctx = gl::context();
-	ctx->setCurrentColor( c );
-}
-
-void texCoord( float s, float t )
-{
-	auto ctx = gl::context();
-	ctx->immediate().texCoord( s, t );
-}
-
-void texCoord( float s, float t, float r )
-{
-	auto ctx = gl::context();
-	ctx->immediate().texCoord( s, t, r );
-}
-
-void texCoord( float s, float t, float r, float q )
-{
-	auto ctx = gl::context();
-	ctx->immediate().texCoord( s, t, r, q );
-}
-
-void texCoord( const ci::vec2 &v )
-{
-	auto ctx = gl::context();
-	ctx->immediate().texCoord( v.x, v.y );
-}
-
-void texCoord( const ci::vec3 &v )
-{
-	auto ctx = gl::context();
-	ctx->immediate().texCoord( v );
-}
-
-void texCoord( const ci::vec4 &v )
-{
-	auto ctx = gl::context();
-	ctx->immediate().texCoord( v );
-}
-
-void vertex( float x, float y )
-{
-	auto ctx = gl::context();
-	ctx->immediate().vertex( vec4( x, y, 0, 1 ), ctx->getCurrentColor() );
-}
-
-void vertex( float x, float y, float z )
-{
-	auto ctx = gl::context();
-	ctx->immediate().vertex( vec4( x, y, z, 1 ), ctx->getCurrentColor() );
-}
-
-void vertex( float x, float y, float z, float w )
-{
-	auto ctx = gl::context();
-	ctx->immediate().vertex( vec4( x, y, z, w ), ctx->getCurrentColor() );
-}
-
-void vertex( const ci::vec2 &v )
-{
-	auto ctx = gl::context();
-	ctx->immediate().vertex( vec4( v.x, v.y, 0, 1 ), ctx->getCurrentColor() );
-}
-
-void vertex( const ci::vec3 &v )
-{
-	auto ctx = gl::context();
-	ctx->immediate().vertex( vec4( v.x, v.y, v.z, 1 ), ctx->getCurrentColor() );
-}
-
-void vertex( const ci::vec4 &v )
-{
-	auto ctx = gl::context();
-	ctx->immediate().vertex( v, ctx->getCurrentColor() );
-}
-
-#if ! defined( CINDER_GL_ES )
-void polygonMode( GLenum face, GLenum mode )
-{
-	auto ctx = gl::context();
-	ctx->polygonMode( face, mode );
-}
-
-void enableWireframe()
-{
-	gl::polygonMode( GL_FRONT_AND_BACK, GL_LINE );
-}
-
-void disableWireframe()
-{
-	gl::polygonMode( GL_FRONT_AND_BACK, GL_FILL );
-}
-
-bool isWireframeEnabled()
-{
-	auto ctx = gl::context();
-	return ctx->getPolygonMode( GL_FRONT_AND_BACK ) == GL_LINE;
-}
-
-#endif // ! defined( CINDER_GL_ES )
-
-void lineWidth( float width )
-{
-	gl::context()->lineWidth( width );
-}
-
-#if ! defined( CINDER_GL_ES )
-
-void pointSize( float size )
-{
-	glPointSize( size );
-}
-
-#endif // ! defined( CINDER_GL_ES )
 
 void draw( const VboMeshRef& mesh )
 {
@@ -824,147 +51,6 @@ void draw( const VboMeshRef& mesh )
 	ctx->popVao();
 }
 
-//////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Vertex Attributes
-void vertexAttribPointer( GLuint index, GLint size, GLenum type, GLboolean normalized, GLsizei stride, const GLvoid *pointer )
-{
-	context()->vertexAttribPointer( index, size, type, normalized, stride, pointer );
-}
-
-#if ! defined( CINDER_GL_ES )
-void vertexAttribIPointer( GLuint index, GLint size, GLenum type, GLsizei stride, const GLvoid *pointer )
-{
-	context()->vertexAttribIPointer( index, size, type, stride, pointer );
-}
-#endif
-
-void enableVertexAttribArray( GLuint index )
-{
-	context()->enableVertexAttribArray( index );
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Buffers
-void vertexAttrib1f( GLuint index, float v0 )
-{
-	context()->vertexAttrib1f( index, v0 );
-}
-
-void vertexAttrib2f( GLuint index, float v0, float v1 )
-{
-	context()->vertexAttrib2f( index, v0, v1 );
-}
-
-void vertexAttrib3f( GLuint index, float v0, float v1, float v2 )
-{
-	context()->vertexAttrib3f( index, v0, v1, v2 );
-}
-
-void vertexAttrib4f( GLuint index, float v0, float v1, float v2, float v3 )
-{
-	context()->vertexAttrib4f( index, v0, v1, v2, v3 );
-}
-
-void bindBuffer( const BufferObjRef &buffer )
-{
-	context()->bindBuffer( buffer->getTarget(), buffer->getId() );
-}
-
-#if ! defined( CINDER_GL_ES_2 )
-void readBuffer( GLenum src )
-{
-	glReadBuffer( src );
-}
-
-void drawBuffers( GLsizei num, const GLenum *bufs )
-{
-	glDrawBuffers( num, bufs );
-}
-#endif
-
-#if ! defined( CINDER_GL_ES )
-void drawBuffer( GLenum dst )
-{
-	glDrawBuffer( dst );
-}
-#endif
-
-void readPixels( GLint x, GLint y, GLsizei width, GLsizei height, GLenum format, GLenum type, GLvoid *data )
-{
-	glReadPixels( x, y, width, height, format, type, data );
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////
-// toGL conversion functions
-GLenum toGl( geom::Primitive prim )
-{
-	switch( prim ) {
-		case geom::Primitive::LINES:
-			return GL_LINES;
-		break;
-		case geom::Primitive::LINE_STRIP:
-			return GL_LINE_STRIP;
-		break;
-		case geom::Primitive::TRIANGLES:
-			return GL_TRIANGLES;
-		break;
-		case geom::Primitive::TRIANGLE_STRIP:
-			return GL_TRIANGLE_STRIP;
-		break;
-		case geom::Primitive::TRIANGLE_FAN:
-			return GL_TRIANGLE_FAN;
-		default:
-			return 0; // no clear right choice here
-	}
-}
-
-geom::Primitive toGeomPrimitive( GLenum prim )
-{
-	switch( prim ) {
-		case GL_LINES:
-			return geom::Primitive::LINES;
-		break;
-		case GL_LINE_STRIP:
-			return geom::Primitive::LINE_STRIP;
-		break;
-		case GL_TRIANGLES:
-			return geom::Primitive::TRIANGLES;
-		break;
-		case GL_TRIANGLE_STRIP:
-			return geom::Primitive::TRIANGLE_STRIP;
-		break;
-		case GL_TRIANGLE_FAN:
-			return geom::Primitive::TRIANGLE_FAN;
-		default:
-			return geom::Primitive( 65535 ); // no clear right choice here
-	}
-}
-
-std::string uniformSemanticToString( UniformSemantic uniformSemantic )
-{
-	switch( uniformSemantic ) {
-		case UNIFORM_MODEL_MATRIX: return "UNIFORM_MODEL_MATRIX";
-		case UNIFORM_MODEL_MATRIX_INVERSE: return "UNIFORM_MODEL_MATRIX_INVERSE";
-		case UNIFORM_MODEL_MATRIX_INVERSE_TRANSPOSE: return "UNIFORM_MODEL_MATRIX_INVERSE_TRANSPOSE";
-		case UNIFORM_VIEW_MATRIX: return "UNIFORM_VIEW_MATRIX";
-		case UNIFORM_VIEW_MATRIX_INVERSE: return "UNIFORM_VIEW_MATRIX_INVERSE";
-		case UNIFORM_MODEL_VIEW: return "UNIFORM_MODEL_VIEW";
-		case UNIFORM_MODEL_VIEW_INVERSE: return "UNIFORM_MODEL_VIEW_INVERSE";
-		case UNIFORM_MODEL_VIEW_INVERSE_TRANSPOSE: return "UNIFORM_MODEL_VIEW_INVERSE_TRANSPOSE";
-		case UNIFORM_MODEL_VIEW_PROJECTION: return "UNIFORM_MODEL_VIEW_PROJECTION";
-		case UNIFORM_MODEL_VIEW_PROJECTION_INVERSE: return "UNIFORM_MODEL_VIEW_PROJECTION_INVERSE";
-		case UNIFORM_PROJECTION_MATRIX: return "UNIFORM_PROJECTION_MATRIX";
-		case UNIFORM_PROJECTION_MATRIX_INVERSE: return "UNIFORM_PROJECTION_MATRIX_INVERSE";
-		case UNIFORM_NORMAL_MATRIX: return "UNIFORM_NORMAL_MATRIX";
-		case UNIFORM_VIEWPORT_MATRIX: return "UNIFORM_VIEWPORT_MATRIX";
-		case UNIFORM_WINDOW_SIZE: return "UNIFORM_WINDOW_SIZE";
-		case UNIFORM_ELAPSED_SECONDS: return "UNIFORM_ELAPSED_SECONDS";
-		default: return "";
-	}
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Draw*
 void drawArrays( GLenum mode, GLint first, GLsizei count )
 {
 	context()->drawArrays( mode, first, count );
@@ -976,6 +62,7 @@ void drawElements( GLenum mode, GLsizei count, GLenum type, const GLvoid *indice
 }
 
 namespace {
+
 std::array<vec3, 8> getCubePoints( const vec3 &c, const vec3 &size )
 {
 	vec3 s = size * 0.5f;																			// origin points
@@ -1098,7 +185,8 @@ void drawCubeImpl( const vec3 &c, const vec3 &size, bool faceColors )
 	ctx->drawElements( GL_TRIANGLES, 36, GL_UNSIGNED_BYTE, 0 );
 	ctx->popVao();
 }
-}
+
+} // anonymous namespace
 
 void drawCube( const vec3 &center, const vec3 &size )
 {
@@ -1115,8 +203,8 @@ void drawStrokedCube( const vec3 &center, const vec3 &size )
 	auto vertices = getCubePoints( center, size );
 	
 	static std::array<GLubyte, 24> indices = { 0, 1, 1, 2, 2, 3, 3, 0,		// right side connection
-												4, 5, 5, 6, 6, 7, 7, 4,		// left side connection
-												0, 4, 1, 5, 2, 6, 3, 7 };	// right to left connections
+											   4, 5, 5, 6, 6, 7, 7, 4,		// left side connection
+											   0, 4, 1, 5, 2, 6, 3, 7 };	// right to left connections
 	
 	auto ctx = ci::gl::context();
 	const GlslProg* curGlslProg = ctx->getGlslProg();
@@ -1938,7 +1026,7 @@ void drawVector( const vec3& start, const vec3& end, float headLength, float hea
 	vec3 ori = end - normalize( dir ) * headLength;
 	gl::draw( geom::Cone().base( headRadius ).height( headLength ).origin( ori ).direction( dir ) );
 }
-	
+
 namespace {
 void drawStringHelper( const std::string &str, const vec2 &pos, const ColorA &color, Font font, int justification )
 {
@@ -1985,387 +1073,6 @@ void drawStringCentered( const std::string &str, const vec2 &pos, const ColorA &
 void drawStringRight( const std::string &str, const vec2 &pos, const ColorA &color, Font font )
 {
 	drawStringHelper( str, pos, color, font, 1 );
-}
-
-GLenum getError()
-{
-	return glGetError();
-}
-
-std::string getErrorString( GLenum err )
-{
-	switch( err ) {
-		case GL_NO_ERROR:
-			return "GL_NO_ERROR";
-		case GL_INVALID_ENUM:
-			return "GL_INVALID_ENUM";
-		case GL_INVALID_VALUE:
-			return "GL_INVALID_VALUE";
-		case GL_INVALID_OPERATION:
-			return "GL_INVALID_OPERATION";
-		case GL_INVALID_FRAMEBUFFER_OPERATION:
-			return "GL_INVALID_FRAMEBUFFER_OPERATION";
-		case GL_OUT_OF_MEMORY:
-			return "GL_OUT_OF_MEMORY";
-		default:
-			return "";
-	}
-}
-
-void checkError()
-{
-	GLenum errorFlag = getError();
-	if( errorFlag != GL_NO_ERROR ) {
-		CI_LOG_E( "glGetError flag set: " << getErrorString( errorFlag ) );
-		CI_ASSERT( 0 );
-	}
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedVao
-ScopedVao::ScopedVao( Vao *vao )
-	: mCtx( gl::context() )
-{
-	mCtx->pushVao( vao );
-}
-
-ScopedVao::ScopedVao( VaoRef &vao )
-	: mCtx( gl::context() )
-{
-	mCtx->pushVao( vao );
-}
-
-ScopedVao::~ScopedVao()
-{
-	mCtx->popVao();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedBuffer
-ScopedBuffer::ScopedBuffer( const BufferObjRef &bufferObj )
-	: mCtx( gl::context() ), mTarget( bufferObj->getTarget() )
-{
-	mCtx->pushBufferBinding( mTarget, bufferObj->getId() );
-}
-
-ScopedBuffer::ScopedBuffer( GLenum target, GLuint id )
-		: mCtx( gl::context() ), mTarget( target )
-{
-	mCtx->pushBufferBinding( target, id );
-}
-
-ScopedBuffer::~ScopedBuffer()
-{
-	mCtx->popBufferBinding( mTarget );
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedState
-ScopedState::ScopedState( GLenum cap, GLboolean value )
-	: mCtx( gl::context() ), mCap( cap )
-{
-	mCtx->pushBoolState( cap, value );
-}
-
-ScopedState::~ScopedState() {
-	mCtx->popBoolState( mCap );
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedColor
-ScopedColor::ScopedColor()
-	: mCtx( gl::context() )
-{
-	mColor = mCtx->getCurrentColor();
-}
-
-ScopedColor::ScopedColor( const ColorAf &color )
-	: mCtx( gl::context() )
-{
-	mColor = mCtx->getCurrentColor();
-	mCtx->setCurrentColor( color );
-}
-
-ScopedColor::~ScopedColor()
-{
-	mCtx->setCurrentColor( mColor );
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedBlend
-ScopedBlend::ScopedBlend( GLboolean enable )
-	: mCtx( gl::context() ), mSaveFactors( false )
-{
-	mCtx->pushBoolState( GL_BLEND, enable );
-}
-
-//! Parallels glBlendFunc(), implicitly enables blending
-ScopedBlend::ScopedBlend( GLenum sfactor, GLenum dfactor )
-	: mCtx( gl::context() ), mSaveFactors( true )
-{
-	mCtx->pushBoolState( GL_BLEND, GL_TRUE );
-	mCtx->pushBlendFuncSeparate( sfactor, dfactor, sfactor, dfactor );
-}
-
-//! Parallels glBlendFuncSeparate(), implicitly enables blending
-ScopedBlend::ScopedBlend( GLenum srcRGB, GLenum dstRGB, GLenum srcAlpha, GLenum dstAlpha )
-	: mCtx( gl::context() ), mSaveFactors( true )
-{
-	mCtx->pushBoolState( GL_BLEND, GL_TRUE );
-	mCtx->pushBlendFuncSeparate( srcRGB, dstRGB, srcAlpha, dstAlpha );
-}
-
-ScopedBlend::~ScopedBlend()
-{
-	mCtx->popBoolState( GL_BLEND );
-	if( mSaveFactors )
-		mCtx->popBlendFuncSeparate();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedGlslProg
-ScopedGlslProg::ScopedGlslProg( GlslProgRef& prog )
-	: mCtx( gl::context() )
-{
-	mCtx->pushGlslProg( prog.get() );
-}
-
-ScopedGlslProg::ScopedGlslProg( const std::shared_ptr<const GlslProg> &prog )
-	: mCtx( gl::context() )
-{
-	mCtx->pushGlslProg( std::const_pointer_cast<GlslProg>( prog ).get() );
-}
-
-ScopedGlslProg::ScopedGlslProg( const GlslProg* prog )
-	: mCtx( gl::context() )
-{
-	mCtx->pushGlslProg( const_cast<GlslProg*>( prog ) );
-}
-
-ScopedGlslProg::~ScopedGlslProg()
-{
-	mCtx->popGlslProg();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedFramebuffer
-ScopedFramebuffer::ScopedFramebuffer( const FboRef &fbo, GLenum target )
-	: mCtx( gl::context() ), mTarget( target )
-{
-	mCtx->pushFramebuffer( fbo, target );
-}
-
-ScopedFramebuffer::ScopedFramebuffer( GLenum target, GLuint framebufferId )
-	: mCtx( gl::context() ), mTarget( target )
-{
-	mCtx->pushFramebuffer( target, framebufferId );
-}
-
-ScopedFramebuffer::~ScopedFramebuffer()
-{	
-#if ! defined( SUPPORTS_FBO_MULTISAMPLING )
-	mCtx->popFramebuffer( GL_FRAMEBUFFER );
-#else
-	if( mTarget == GL_FRAMEBUFFER || mTarget == GL_READ_FRAMEBUFFER )
-		mCtx->popFramebuffer( GL_READ_FRAMEBUFFER );
-	if( mTarget == GL_FRAMEBUFFER || mTarget == GL_DRAW_FRAMEBUFFER )
-		mCtx->popFramebuffer( GL_DRAW_FRAMEBUFFER );
-#endif
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedActiveTexture
-ScopedActiveTexture::ScopedActiveTexture( uint8_t textureUnit )
-	: mCtx( gl::context() )
-{
-	mCtx->pushActiveTexture( textureUnit );
-}
-	
-ScopedActiveTexture::~ScopedActiveTexture()
-{
-	mCtx->popActiveTexture();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedTextureBind
-ScopedTextureBind::ScopedTextureBind( GLenum target, GLuint textureId )
-	: mCtx( gl::context() ), mTarget( target )
-{
-	mTextureUnit = mCtx->getActiveTexture();
-	mCtx->pushTextureBinding( mTarget, textureId, mTextureUnit );
-}
-
-ScopedTextureBind::ScopedTextureBind( GLenum target, GLuint textureId, uint8_t textureUnit )
-	: mCtx( gl::context() ), mTarget( target ), mTextureUnit( textureUnit )
-{
-	mCtx->pushTextureBinding( mTarget, textureId, mTextureUnit );
-}
-
-ScopedTextureBind::ScopedTextureBind( const TextureBaseRef &texture )
-	: mCtx( gl::context() ), mTarget( texture->getTarget() )
-{
-	mTextureUnit = mCtx->getActiveTexture();
-	mCtx->pushTextureBinding( mTarget, texture->getId(), mTextureUnit );
-}
-
-ScopedTextureBind::ScopedTextureBind( const TextureBaseRef &texture, uint8_t textureUnit )
-	: mCtx( gl::context() ), mTarget( texture->getTarget() ), mTextureUnit( textureUnit )
-{
-	mCtx->pushTextureBinding( mTarget, texture->getId(), mTextureUnit );
-}
-
-// ----------------------------------------------------------------
-// These overloads are to alleviate a VS2013 bug where it cannot deduce
-// the correct constructor when a TextureBaseRef subclass is passed in
-ScopedTextureBind::ScopedTextureBind( const Texture2dRef &texture, uint8_t textureUnit )
-	: mCtx( gl::context() ), mTarget( texture->getTarget() ), mTextureUnit( textureUnit )
-{
-	mCtx->pushTextureBinding( mTarget, texture->getId(), mTextureUnit );
-}
-
-#if ! defined( CINDER_GL_ES_2 )
-ScopedTextureBind::ScopedTextureBind( const Texture3dRef &texture, uint8_t textureUnit )
-	: mCtx( gl::context() ), mTarget( texture->getTarget() ), mTextureUnit( textureUnit )
-{
-	mCtx->pushTextureBinding( mTarget, texture->getId(), mTextureUnit );
-}
-#endif // ! defined( CINDER_GL_ES_2 )
-
-ScopedTextureBind::ScopedTextureBind( const TextureCubeMapRef &texture, uint8_t textureUnit )
-	: mCtx( gl::context() ), mTarget( texture->getTarget() ), mTextureUnit( textureUnit )
-{
-	mCtx->pushTextureBinding( mTarget, texture->getId(), mTextureUnit );
-}
-// ----------------------------------------------------------------
-
-ScopedTextureBind::~ScopedTextureBind()
-{
-	mCtx->popTextureBinding( mTarget, mTextureUnit );
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedScissor
-ScopedScissor::ScopedScissor( const ivec2 &lowerLeftPosition, const ivec2 &dimension )
-	: mCtx( gl::context() )
-{
-	mCtx->pushBoolState( GL_SCISSOR_TEST, GL_TRUE );
-	mCtx->pushScissor( std::pair<ivec2, ivec2>( lowerLeftPosition, dimension ) );
-}
-
-ScopedScissor::ScopedScissor( int lowerLeftX, int lowerLeftY, int width, int height )
-	: mCtx( gl::context() )
-{
-	mCtx->pushBoolState( GL_SCISSOR_TEST, GL_TRUE );
-	mCtx->pushScissor( std::pair<ivec2, ivec2>( ivec2( lowerLeftX, lowerLeftY ), ivec2( width, height ) ) );		
-}
-	
-ScopedScissor::~ScopedScissor()
-{
-	mCtx->popBoolState( GL_SCISSOR_TEST );
-	mCtx->popScissor();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedFaceCulling
-ScopedFaceCulling::ScopedFaceCulling( bool cull )
-	: mCtx( gl::context() ), mSaveFace( false )
-{
-	mCtx->pushBoolState( GL_CULL_FACE, cull );
-}
-
-ScopedFaceCulling::ScopedFaceCulling( bool cull, GLenum face )
-	: mCtx( gl::context() ), mSaveFace( true )
-{
-	mCtx->pushBoolState( GL_CULL_FACE, cull );
-	mCtx->pushCullFace( face );
-}
-
-ScopedFaceCulling::~ScopedFaceCulling()
-{
-	mCtx->popBoolState( GL_CULL_FACE );
-	if( mSaveFace )
-		mCtx->popCullFace();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedRenderbuffer
-ScopedRenderbuffer::ScopedRenderbuffer( const RenderbufferRef &rb )
-	: mCtx( gl::context() )
-{
-	mCtx->pushRenderbufferBinding( GL_RENDERBUFFER, rb->getId() );
-}
-
-ScopedRenderbuffer::ScopedRenderbuffer( GLenum target, GLuint id )
-	: mCtx( gl::context() )
-{
-	// this is the only legal value currently
-	CI_ASSERT( target == GL_RENDERBUFFER );
-	mCtx->pushRenderbufferBinding( target, id );
-}
-
-ScopedRenderbuffer::~ScopedRenderbuffer()
-{
-	mCtx->popRenderbufferBinding( GL_RENDERBUFFER );
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedViewport
-ScopedViewport::ScopedViewport( const ivec2 &lowerLeftPosition, const ivec2 &dimension )
-	: mCtx( gl::context() )
-{
-	mCtx->pushViewport( std::pair<ivec2, ivec2>( lowerLeftPosition, dimension ) );
-}
-
-ScopedViewport::ScopedViewport( int lowerLeftX, int lowerLeftY, int width, int height )
-	: mCtx( gl::context() )
-{
-	mCtx->pushViewport( std::pair<ivec2, ivec2>( ivec2( lowerLeftX, lowerLeftY ), ivec2( width, height ) ) );		
-}
-	
-ScopedViewport::~ScopedViewport()
-{
-	mCtx->popViewport();
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedLineWidth
-ScopedLineWidth::ScopedLineWidth( float width )
-	: mCtx( gl::context() )
-{
-	mCtx->pushLineWidth( width );
-}
-
-ScopedLineWidth::~ScopedLineWidth()
-{
-	mCtx->popLineWidth();
-}
-
-#if ! defined( CINDER_GL_ES )
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedPolygonMode
-ScopedPolygonMode::ScopedPolygonMode( GLenum mode )
-	: mCtx( gl::context() )
-{
-	mCtx->pushPolygonMode( GL_FRONT_AND_BACK, mode );
-}
-
-ScopedPolygonMode::~ScopedPolygonMode()
-{
-	mCtx->popPolygonMode( GL_FRONT_AND_BACK );
-}
-#endif
-
-///////////////////////////////////////////////////////////////////////////////////////////
-// ScopedFrontFace
-ScopedFrontFace::ScopedFrontFace( GLenum mode )
-	: mCtx( gl::context() )
-{
-	mCtx->pushFrontFace( mode );
-}
-
-ScopedFrontFace::~ScopedFrontFace()
-{
-	mCtx->popFrontFace();
 }
 
 } } // namespace cinder::gl
